test: adapt to pass under newer WP
Test was failing due to changes in newer WP that make
createNewPost() timeout.

createNewPost comes from @wordpress/e2e-test-utils, which is now
deprecated in favour of @wordpress/e2e-test-utils-playwright.

https://www.npmjs.com/package/@wordpress/e2e-test-utils?activeTab=versions

We'd have to migrate to @wordpress/e2e-test-utils-playwright to fix
this more cleanly.

But migrating to @wordpress/e2e-test-utils-playwright requires
significant changes including likely:

- Node.js version bump
- React upgrade
- Bump of all `wordpress/` packages.
- Changing testing infrastructure to use playwright (requires rewriting
all tests)
- Possibly dropping pptr-testing-library.
- Likely updates to linting for Node.js bump

I opted not to do that here because our goal is simply to unblock the CI
pipeline so anyone can deploy 'tested up to' changes without local
SVN creds.

We'll likely have to do the above to continue testing future versions
of WP, and we already have a template to follow from the similar upgrade
I did in Genesis Blocks:

https://github.com/studiopress/genesis-blocks/pull/353

But for now this should be enough to unblock the pipeline and retain
the test without having to skip it entirely.